[Refactor] AlbumAndTagListView VC에서 분리
